test: move binary integration tests to tests/cli.rs (#374)
Move the 3 binary-invoking tests from src/main.rs inline test module to
a proper integration test file. This fixes binary lookup for custom
build-dir configurations, new build-dir layouts, and other Cargo build
directory changes.

Cargo only sets CARGO_BIN_EXE_<name> for integration tests, so using
cargo_bin_cmd!() in tests/cli.rs resolves the binary path at compile
time via env!() rather than relying on fragile fallback path lookups.
